requirements.txt
tensorflow
google-cloud-dataflow
numpy
Pillow
scikit-image
apache-beam[gcp]
pandas
sklearn
scipy